Moved putchar and puts Moved to stdio.hpp and their own source files
